dataset: cleanup datasets that hit the memcap while loading
Datasets that hit the memcap limit need to be discarded if the memcap is
hit or otherwise the datasets are still loaded with partial data while
the signature is not loaded due to the memcap error. This moves the
check prior to the final cleanup. This also adds the check for datarep
sets a well due to the move to the generic dataset code section.

Ticket: #6678