[AWN-314905] Update caps used by Suricata
The capabilities assigned when dropping
priveleges are not sufficient to run XDP
applications.

Additionally, the priveleges required differ
depending on Linux version (CAP_BPF was added
in Linux 5.8 to break out of the already
overused CAP_SYS_ADMIN role).

Query the Linux version at run-time in order
to determine which set of capabilities to
use.