output: Add linktype description
Amend the linktype output with the linktype name (when available).

The linktype name is included alongside linktype when `alert.packet` is
enabled. The name is retrieved from pcap_datalink_val_to_name with the
thought being that the pcap library in use can also translate the
linktype value into a descriptive value.