detect/content-inspect: bench tool
Tool to benchmark detection engine content inspection, which is the
inspection of individual groups of content, etc matches for a buffer.

Also add a set of basic tests for the various single pattern matching
implementation.

Output is in csv. To files for the rule based tests. To stdout for the
spm tests.