datasets: fix memuse to include string len
So far, when the data size was passed to the hash table, it was sent as
a sizeof(Struct) which works fine for the other data types as they have
a fixed length but not for the StringType. But, because of the sizeof
construct, the length of a string type dataset was always taken to be 16
Bytes which is only the size of the struct itself. It did not accomodate
the actual size of the string that the StringType holds. Fix this so the
memuse that is used to determine whether memcap was reached also takes
into consideration the size of the actual string.

Bug 3910