SSC-1883 Update tcp_init_flags if first packets in flow arrive late
In a TCP flow, suricata may see the 2nd packet from the client or
server before the 1st packet. This change handles this case, and updates
the `tcp_init_flags` field for the corresponding flow direction(s).

If the 1st packet in either directions is seen late, ensure that the
`STREAMTCP_FLAG_MIDSTREAM` flag is cleared if it no longer applies.

Testing:
- verified that an non-midstream flow observation is generated if the
  1st packets from the client and server aren't seen by suricata until
  after the 2nd packets. This was done by creating a custom pcap file
  using `editcap` and `mergecap`