flow: exit emergency mode based on current memuse
Flow Manager will exit the emergency mode when the spare pool has a
certain percentage of flows defined by flow_config.emergency_recovery
w.r.t prealloc setting. This can lead to a flip flop condition of
exiting and entering emergency mode in deployments with heavy traffic
and low memcap and prealloc settings.
Use the current in memory flows to determine when should emergency mode
be over.
This comes with an obvious change in the behavior of the aforementioned
deployments in the form of FM never being able to come out of emergency
mode but that is better than once coming out and then quickly jumping
into emergency back again.

e.g. for a prealloc setting of 10k
emergency_recovery setting of 30
in memory flows of 20k

Current scenario: will declare emergency mode over as soon as it has >3k
flows in the spare pool irrespective of current in memory flows.

After this patch: will declare emergency mode over only when it has >6k
flows in the spare pool.