http2/doh: add required padding
The newer base64 crate would reject any input that cannot be translated
to perfect 4 byte base64 blocks. Add the required padding to the input
in case the input is not a multiple of 4. This change makes sure the
parsing is interpreted the same way as the older crate.

Related to Task 7219