github/pr-template: update contribution checks
We have different types of checks, all in a single list.
Adding a title of sorts to each group in the hopes of highlighting
their purposes.

When we add new items to the json schema, many times we don't add their
descriptions, while this would be the perfect time to also do that.