applayer: count flows that errored out
In some cases in the applayer parser, if the respective applayer parser
encountered an error during parsing, the flow counter was not
incremented. This is not ideal as the stat should report the total
number of flows seen for a certain protocol by the parser. Whether such
a flow errored out or not is depicted by other stats.
This removes discrepancies in the stats where in case of many errors,
the reported number of flows could be really less.