detect/http: fix progress for headers keywords
Ticket: 7326

Having a lower progress than the right one, made prefilter
bail out too early, not having found a buffer in the multi-buffer
that mathed the prefilter.

Hence, we got false negatives, if http.request_header or
http.response_header was used as fast pattern, and if the request
or response came in multiple packets, and the first of these packets
did not have enough data (like only http request line),
and the next packets did have the matching data.