make: install-headers: rust-bindings.h
rust-bindings.h was not being installed with rust-bindings.h, and its
now pulled in by a header used for plugin support, so make sure its
installed.

We first attempt to install the "dist" version if exists, otherwise
install the "gen" one. Also install the "gen" even if the "dist" one
exists, as its going to be newer.