app-layer: fix smb/dcerpc test
Test AppLayerTest08 has an invalid DCERPC request in many ways.
- header type is invalid
- fragment length is invalid

With the new changes of using the applayer::incomplete API in dcerpc
parser, this test failed due to dcerpc parser asking for more data which
does not lead to any condition that would ask for disabling applayer.
Applayer can be disabled from two code paths:
1. StreamTcpDisableAppLayer
2. StreamTcpSetSessionNoReassemblyFlag

These fns are called in the cases of:
- applayer parser not requiring inspection
- applayer parser asking to not reassemble data
- stream depth having been reached
- errors encountered by applayer during parsing

Since none of the these conditions are the aim of the given test, adjust
the byte array to actually contain valid data that goes through all the
conditions as expected in the FAIL/PASS macros.