rust: integrate bindgen to generate rust bindings to C
Add a minimal integration of bindgen to the build.

Bindgen is integrated at compile time with a "build.rs" file that for
now only generates AppLayerEventType Rust bindings.

This required some refactoring of the C so app-layer-events.h did not
also include "rust.h", which causes issues for bindgen, probably
related to circular references.

AppLayerEventType was chosen as the first step as its an argument type
some app-layer functions that we may want to use bindgen to export
Rust, and one of the requirements of bindgen might be that C functions
should only use datatypes defined in C, and not Rust. Following such a
rule also prevents circular dependencies between Rust and C code.

Bindgen generates the bindings in a file named "bindings.rs" in the
target/ direction, which "sys.rs" will statically "include" at
compiling name, making these bindings available under package
"crate::sys".

"build.rs" had to be placed in the non-standard location of
"src/build.rs" (its usually alongside Cargo.toml) to satisfy the
out-of-tree build requirements of distcheck.

Note that bindgen is also available as a command line tool which could
be used instead of integrating this at compile time, however the tool
requires a newer version of Rust than our MSRV, and may present
additional issues with respect to autoconf and distcheck.

Ticket: #7341